pyproject.toml
[project]
name = "joinly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "dbrockmann", email = "dominik@joinly.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "faster-whisper>=1.1.1",
    "fastmcp>=2.3.3",
    "kokoro-onnx>=0.4.8",
    "langchain-mcp-adapters>=0.0.3",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.1",
    "mcp>=1.7.1",
    "numpy>=2.2.4",
    "playwright>=1.51.0",
    "python-dotenv>=1.1.0",
    "semchunk>=3.2.1",
    "silero-vad-lite>=0.2.1",
]

[project.scripts]
joinly = "joinly.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.18",
    "jiwer>=3.1.0",
    "pre-commit>=4.2.0",
    "pyright>=1.1.399",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "rich>=14.0.0",
    "ruff>=0.11.6",
]

[tool.hatch.build.targets.wheel]
packages = ["joinly"]
